+++ include/Stack.h
@@ -6,9 +6,9 @@ class Stack {
 public:
   virtual ~Stack() {}
   virtual bool isEmpty() = 0;
-  virtual void push(const int data) = 0;
-  virtual int pop() = 0;
-  virtual int top() = 0;
+  virtual void push(const unsigned long long data) = 0;
+  virtual unsigned long long pop() = 0;
+  virtual unsigned long long top() = 0;
 };
 
 class LinkedStack : public Stack {
@@ -21,13 +21,13 @@ public:
   virtual bool isEmpty();
   
   // Adds data to the top of the linked stack
-  virtual void push(const int data);
+  virtual void push(const unsigned long long data);
 
   // Removes and returns the value stored at top
-  virtual int pop();
+  virtual unsigned long long pop();
 
   // Returns the value stored at top
-  virtual int top();
+  virtual unsigned long long top();
 
   // Destructor
   virtual ~LinkedStack() {}
